+++ audio_visualizer/src/androidTest/java/com/miller198/audiovisualizer/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.miller198.audio_visualizer
+package com.miller198.audiovisualizer
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
