+++ audio_visualizer/src/test/java/com/miller198/audiovisualizer/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.miller198.audio_visualizer
+package com.miller198.audiovisualizer
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
